+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.1.1] - Added delayed menu items
+
+* Added the ability to call menu items after a delay either via dedicated menu item for the delayed action or via notification from another module.
+    - Example uses: refresh the page after X seconds, or turn off the monitor after X seconds.
+
 ## [0.1.0] - Added Notification Control Options & Multi-Instance
 
 * Added the ability to control the menu itself via notifications from other modules. See [README](readme.md#controlling-the-menu-from-another-module) for additional details.